Remove deferral of core JavaScript
In OnTopic 5.1.0, as part of some performance improvements, we deferred all JavaScript. This potentially improves Lighthouse rankings by avoiding blocking scripts. In doing so, however, it also causes layout shifts in the interface as, notably, the Popper.js tooltips get rendered as inline text until the JavaScript is loaded and initialized. As such, the actual performance perception decreases—as might the Lighthouse ranking.

(Note: As the OnTopic Editor interface isn't expected to be publicly exposed, Google rankings aren't important to it—but we were using it in OnTopic 5.1.0 as a quick measure of low-hanging fruit for improving performance.)