Check for whitespace, not just empty, for `GetInteger()`
This isn't an expected scenario, but sometimes whitespace is used to ensure a previous value is overridden in the editor. Whitespace values cannot be converted to an integer anymore than an empty value can be. To help detect these scenarios, without relying on `TryParse()`, I'm using `IsNullOrWhiteSpace()` instead of `IsNullOrEmpty()`.