Marked unmapped properties as `[DisableMapping]`
The `StyleSheets` and `Scripts` collections are populated by the constructors of derived types, and don't map to associations on the source `AttributeDescriptor`. The `IsValueRequired` is a calculated property based on other properties; while those dependency properties are mapped, the `IsValueRequired` property is not. As none of these had setters, they would have been passed over by the `TopicMappingService`, but that to happen much earlier in the process. This should have a small performance benefit.