Updated JSHint to minor version
Updated the JSHint plugin from 2.12.0 to 2.12.4. This was previously updated to 2.12.2 in the main `OnTopic.Editor.AspNetCore` project, but was missed for `OnTopic.Editor.AspNetCore.Attributes`; this picks that up, and updates both to the latest patch. This is a build dependency and won't impact consumers of this package directly.