Updated SASS to major version
Updated the Gulp SASS plugin from 4.1.0 to 5.1.0, and also referenced the currently used `node-sass` compiler. This was previously updated for the main `OnTopic.Editor.AspNetCore` project, but was missed for `OnTopic.Editor.AspNetCore.Attributes`; this picks that up. This is a build dependency and won't impact consumers of this package directly. As part of this, updated the `gulpFile` to specify the SASS compiler (`node-sass`), as required by the 5.x release. The compiler will be updated for both projects in a subsequent commit.