Revert: Renamed `AttributeDictionary` to `AttributeValueDictionary`
The latest patch renamed `AttributeValueDictionary` back to `AttributeDictionary`, thus reverting a large part of the previous patch (b243853). It seems this decision was premature, though, as we can mitigate the naming conflicts without renaming the class, since the out-of-the-box `ViewFeatures` namespace is almost exclusively used in views, and the `AttributeDictionary` is almost exclusively used in view models. This effectively undoes 1ca4eb4 and OnTopicCMS/OnTopic-Library#107.